+++ lib/fortnox/api/repositories/article.rb
@@ -6,10 +6,13 @@ module Fortnox
   module API
     module Repository
       class Article < Fortnox::API::Repository::Base
-
         MODEL = Fortnox::API::Model::Article
         MAPPER = Fortnox::API::Mapper::Article
         URI = '/articles/'.freeze
+
+        def initialize
+          super(MODEL)
+        end
       end
     end
   end
