+++ lib/fortnox/api/repositories/base.rb
@@ -1,23 +1,90 @@
-require "fortnox/api/base"
+require 'httparty'
+require 'fortnox/api/request_handling'
 require "fortnox/api/repositories/base/loaders"
 require "fortnox/api/repositories/base/savers"
 
 module Fortnox
   module API
     module Repository
-      class Base < Fortnox::API::Base
-
+      class Base
+        include HTTParty
+        include Fortnox::API::RequestHandling
         include Loaders
         include Savers
 
+        HTTParty::Parser::SupportedFormats[ "text/html" ] = :json
+
+        DEFAULT_HEADERS = {
+          'Content-Type' => 'application/json',
+          'Accept' => 'application/json',
+        }.freeze
+
+        HTTP_METHODS = [ :get, :put, :post, :delete ].freeze
+
+        attr_accessor :headers
         attr_reader :mapper, :keys_filtered_on_save
 
-        def initialize( keys_filtered_on_save: [ :url ], token_store: :default )
-          super()
+        def self.set_headers( headers = {} )
+          self.headers.merge!( headers )
+        end
+
+        HTTP_METHODS.each do |method|
+          define_method method do |*args, token_store:|
+            self.headers['Access-Token'] = get_access_token(token_store)
+            execute do |remote|
+              remote.send( method, *args )
+            end
+          end
+        end
+
+        def initialize( model, keys_filtered_on_save: [ :url ], token_store: :default )
+          self.class.base_uri( get_base_url )
+
+          self.headers = DEFAULT_HEADERS.merge({
+            'Client-Secret' => get_client_secret,
+          })
 
           @keys_filtered_on_save = keys_filtered_on_save
           @token_store = token_store
-          @mapper = Registry[ Mapper::Base.canonical_name_sym( self.class::MODEL )].new
+          @mapper = Registry[ Mapper::Base.canonical_name_sym( model )].new
+        end
+
+        def config
+          Fortnox::API.config
+        end
+
+        def get_access_token( store_name )
+          @access_tokens ||= CircularQueue.new( *load_access_tokens( store_name ))
+          @access_tokens.next
+        end
+
+        def check_access_tokens!( tokens )
+          if tokens == nil or tokens.length.zero?
+            fail MissingConfiguration, "You have not provided any access token(s) in the given store."
+          end
+        end
+
+        def load_access_tokens( store_name )
+          begin
+            tokens = config.token_store.fetch( store_name )
+          rescue KeyError
+            fail MissingConfiguration, "Store #{store_name.inspect} is not present in token store."
+          end
+
+          check_access_tokens!( tokens )
+          tokens
+        end
+
+        def get_base_url
+          base_url = config.base_url
+          fail MissingConfiguration, 'You have to provide a base url.' unless base_url
+          base_url
+        end
+
+        def get_client_secret
+          client_secret = config.client_secret
+          fail MissingConfiguration, 'You have to provide your client secret.' unless client_secret
+          client_secret
         end
 
         private
