+++ lib/fortnox/api/repositories/invoice.rb
@@ -6,10 +6,13 @@ module Fortnox
   module API
     module Repository
       class Invoice < Fortnox::API::Repository::Base
-
         MODEL = Fortnox::API::Model::Invoice
         MAPPER = Fortnox::API::Mapper::Invoice
         URI = '/invoices/'.freeze
+
+        def initialize
+          super(MODEL)
+        end
       end
     end
   end
